postgres_to_es/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
backoff==2.2.1
elasticsearch==8.16.0
uvicorn==0.34.0